requirements.txt
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
darkdetect==0.8.0
Deprecated==1.2.14
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Limiter==3.7.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
h11==0.14.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.1
openalgo==1.0.2
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
Pygments==2.18.0
PyJWT==2.8.0
pyngrok==7.1.6
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
setuptools==70.3.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
zipp==3.19.2
AutoTrader-Web-API-Stocks-Developer